Added integration test for publicmessages
